MCP-92 Update unit tests
